requirements.txt
pandas==2.1.3
pandera==0.18.1
pyyaml==6.0.1
pytest==7.4.3
